seedapp_microservices/user_service/requirements.txt
Django>=3.2
djangorestframework
pymysql
gunicorn
django-cors-headers
pillow


